chore(pay-card): share the card MSW test store across flows
