+++ features/flow/pay-card-details/src/components/CardDetails/CardDetails.native.test.tsx
@@ -1,16 +1,19 @@
 import React, { type PropsWithChildren } from "react";
 import { render, screen, userEvent } from "@testing-library/react-native";
-import { SignedInCardApiProviders, listenToSignedInCardApi } from "../../__tests__/cardApiStore";
+import { cardApiWrapper, listenToCardApi } from "@support/msw-features-flow-pay-card";
 import { CARD_COPY, MORE_COPY, I18nWrapper } from "../../__tests__/i18nWrapper";
+import { signedInCardApiHandlers } from "./signedInCardApi";
 import { CardDetails } from "./CardDetails";
 
-listenToSignedInCardApi();
+listenToCardApi(signedInCardApiHandlers);
+
+const StoreWrapper = cardApiWrapper({ signedIn: true });
 
 function Wrapper({ children }: PropsWithChildren) {
   return (
-    <SignedInCardApiProviders>
+    <StoreWrapper>
       <I18nWrapper>{children}</I18nWrapper>
-    </SignedInCardApiProviders>
+    </StoreWrapper>
   );
 }
