+++ features/flow/pay-card-details/src/components/CardNumbers/CardNumbers.web.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
-import { CardApiStoreProvider, makeCardApiStore } from "../../__tests__/cardApiStore";
+import { CardApiStoreProvider, makeCardApiStore } from "@support/msw-features-flow-pay-card";
 import { CardNumbers } from "./CardNumbers";
 
 describe("CardNumbers (web)", () => {
