+++ support/README.md
@@ -39,6 +39,7 @@ Two kinds of preset:
 | [`jest-features-flow`](./jest-features-flow) | `features/flow/*` — dual web/native jest preset plus Lumen passthrough stubs |
 | [`jest-shared`](./jest-shared) | `shared/*` — flat node preset for logic packages, dual web/native and native-only presets for UI packages |
 | [`lint-rules`](./lint-rules) | monorepo-wide custom ESLint rules |
+| [`msw-features-flow-pay-card`](./msw-features-flow-pay-card) | `features/flow/pay-card-*` — MSW + RTK Query test store and server |
 
 ## Adding a package
