+++ support/msw-features-flow-pay-card/package.json
@@ -1,0 +1,43 @@
+{
+  "name": "@support/msw-features-flow-pay-card",
+  "version": "0.1.0",
+  "private": true,
+  "description": "Shared MSW + RTK Query test harness for Pay Card features/flow packages",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./package.json": "./package.json"
+  },
+  "sideEffects": false,
+  "scripts": {
+    "format": "oxfmt src",
+    "format:check": "oxfmt src --check",
+    "lint": "oxlint src",
+    "lint:ci": "oxlint src --quiet",
+    "lint:fix": "oxfmt src && oxlint src --fix --quiet",
+    "typecheck": "tsc --noEmit"
+  },
+  "dependencies": {
+    "@domain/api-card-management": "workspace:*",
+    "@features/flow-pay-card-auth": "workspace:*",
+    "@reduxjs/toolkit": "catalog:",
+    "@shared/api-services": "workspace:*"
+  },
+  "peerDependencies": {
+    "msw": "catalog:",
+    "react": "catalog:",
+    "react-redux": "catalog:"
+  },
+  "devDependencies": {
+    "@types/jest": "catalog:",
+    "@types/node": "catalog:",
+    "@types/react": "catalog:",
+    "msw": "catalog:",
+    "oxfmt": "catalog:",
+    "oxlint": "catalog:",
+    "react": "catalog:",
+    "react-redux": "catalog:",
+    "typescript": "catalog:"
+  }
+}
