+++ support/msw-features-flow-pay-card/project.json
@@ -1,0 +1,8 @@
+{
+  "name": "@support/msw-features-flow-pay-card",
+  "targets": {
+    "build": {
+      "executor": "nx:noop"
+    }
+  }
+}
